requirements.txt
Keras==2.1.3
tensorflow==1.5.0
h5py==2.7.1
Pillow==4.3.0
coremltools==0.8
